% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFocals.R
\name{findFocals}
\alias{findFocals}
\title{findFocals}
\usage{
findFocals(snps, sample.name, ignore.chr.Y=NULL, min.snps.in.cluster=5, max.extension=6,
fdel.sd.lim=4, fdel.min.num.below.thr=6, fdel.min.pct.below.thr=0.3, fdel.max.pct.above.mean=0.1,
fgain.sd.lim=3.5, fgain.min.num.above.thr=6, fgain.min.pct.above.thr=0.4, fgain.max.pct.below.mean=0.2,
cbs.min.segment.length=2e6, cbs.alpha=0.01, verbose=TRUE)
}
\arguments{
\item{snps}{The SNPs data}

\item{sample.name}{The name of the sample}

\item{ignore.chr.Y}{(defaults to NULL)}

\item{min.snps.in.cluster}{(defaults to 5)}

\item{max.extension}{(defaults to 6)}

\item{fdel.sd.lim}{(defaults to 4)}

\item{fdel.min.num.below.thr}{(defaults to 6)}

\item{fdel.min.pct.below.thr}{(defaults to 0.3)}

\item{fdel.max.pct.above.mean}{(defaults to 0.1)}

\item{fgain.sd.lim}{(defaults to 3.5)}

\item{fgain.min.num.above.thr}{(defaults to 6)}

\item{fgain.min.pct.above.thr}{(defaults to 0.4)}

\item{fgain.max.pct.below.mean}{(defaults to 0.2)}

\item{cbs.min.segment.length}{(defaults to 2e6)}

\item{cbs.alpha}{(defaults to 0.01)}

\item{verbose}{(defaults to TRUE)}
}
\value{
A "FindFocalsResults" object with the results of the analysis. The object is
a list with the results for gains and deletions in different slots.
}
\description{
Find clusters of extrem LRR values and determine if they might be classified
as a focal gain or deletion.
}
\details{
Given a GRanges with SNP data as the one created by loadSNPData, it
processess the LRR values to detect clusters of outliers, that is,
clusters of SNP a number of standard deviations above or below the
mean LRR. Once detected, it will try to extend the detected clusters until
the addition of a new SNP fails to comply with the required thresholds.
}
\examples{



}
