% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFocals.R
\name{plotFocals}
\alias{plotFocals}
\title{plotFocals}
\usage{
plotFocals(find.focal.results,
plot.gain=TRUE, plot.gain.points=FALSE, gain.col="#FFBD07AA", gain.border.col="#FFBD07AA",
plot.del=TRUE, plot.del.points=FALSE, del.col="#00A6EDAA", del.border.col="#00A6EDAA",
plot.mean.lrr=FALSE, mean.lrr.col="#B0FFA5AA", mean.lrr.border.col=NA, mean.lrr.line.col="#FF7C30",
main="Focal Gains and Losses", chromosomes="canonical", zoom=NULL,
lrr.min=-4, lrr.max=2, total.height=1, bottom=0, margin=0.05,
focal.points.cex=0.6, points.cex=0.3, labels.cex=1.5, main.cex=2, axis.cex=1.2, chr.cex=1.5)
}
\arguments{
\item{find.focal.results}{The results from findFocals}

\item{plot.gain}{(defaults to TRUE)}

\item{plot.gain.points}{(defaults to FALSE)}

\item{gain.col}{(defaults to "#FFBD07AA")}

\item{gain.border.col}{(defaults to "#FFBD07AA")}

\item{plot.del}{(defaults to TRUE)}

\item{plot.del.points}{(defaults to FALSE)}

\item{del.col}{(defaults to "#00A6EDAA")}

\item{del.border.col}{(defaults to "#00A6EDAA")}

\item{plot.mean.lrr}{(defaults to FALSE)}

\item{mean.lrr.col}{(defaults to "#B0FFA5AA")}

\item{mean.lrr.border.col}{(defaults to NA)}

\item{mean.lrr.line.col}{(defaults to "#FF7C30")}

\item{main}{(defaults to "Focal Gains and Losses")}

\item{chromosomes}{(defaults to "canonical")}

\item{zoom}{(defaults to NULL)}

\item{lrr.min}{(defaults to -4)}

\item{lrr.max}{(defaults to 2)}

\item{total.height}{(defaults to 1)}

\item{bottom}{(defaults to 0)}

\item{margin}{(defaults to 0.05)}

\item{focal.points.cex}{(defaults to 0.6)}

\item{points.cex}{(defaults to 0.3)}

\item{labels.cex}{(defaults to 1.5)}

\item{main.cex}{(defaults to 2)}

\item{axis.cex}{(defaults to 1.2)}

\item{chr.cex}{(defaults to 1.5)}
}
\value{
Invisibly returns the karyoplot object representing the plot. With it
it is possible to add other elements to the plot using standrad karyoploteR
functions
}
\description{
Plots the SNP array data with the identified focal gains and losses on top.
}
\details{
Creates a plot with the LRR and BAF values along the genome and adds
rectangles showing the identified focal gains and deletions. Optionally,
the SNPs in each of the focal events may be highlighted.
}
\examples{



}
